golang fix

+++ otelcollector/fluent-bit/src/timeseries_volume.go
@@ -110,7 +110,7 @@ func PublishTimeseriesVolume() {
 			if os.Getenv("AZMON_INVALID_CUSTOM_PROMETHEUS_CONFIG") == "true" {
 				isInvalidCustomConfig = 1
 			}
-			invalidCustomConfigMetric.With(prometheus.Labels{"computer":CommonProperties["computer"], "release":CommonProperties["helmreleasename"], "controller_type":CommonProperties["controllertype"]}).Set(isInvalidCustomConfig)
+			invalidCustomConfigMetric.With(prometheus.Labels{"computer":CommonProperties["computer"], "release":CommonProperties["helmreleasename"], "controller_type":CommonProperties["controllertype"]}).Set(float64(isInvalidCustomConfig))
 		
 			lastTickerStart = time.Now()
 		}
